Extract scheduling to its own Cog
